requirements.txt
flask
flask-cors
gunicorn
pymongo[srv]
python-dotenv
google-generativeai
